requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pymysql==1.1.0
gitpython==3.1.40
MarkupSafe==2.1.3
